+++ processor/processor.ts
@@ -1,4 +1,4 @@
-import { InputData, ChartOptions, ChartData } from '../types/types';
+import { InputData, ChartOptions, ChartData, InputPoint } from '../types/types';
 import ChartValidator from '../validator/validator';
 
 class ChartProcessor {
@@ -10,9 +10,13 @@ class ChartProcessor {
         this.options = options;
     }
 
+    getPointValue(point: InputData): number {
+        return typeof point === "number" ? point : point.value;
+    }
+
     calculateTotal(data: InputData[]): number {
         return data.reduce<number>((a, p) => {
-            const value = typeof p === "number" ? p : p.value;
+            const value = this.getPointValue(p);
             return a + value;
         }, 0);
     }
@@ -21,23 +25,18 @@ class ChartProcessor {
         const total = this.calculateTotal(data);
 
         return data.reduce<number>((a, p) => {
-            const value = typeof p === "number" ? p : p.value
+            const value = this.getPointValue(p);
             const label = typeof p === "number" ? p.toString() : (p.label || p.value.toString());
 
-            this.options.max.value = value >= this.options.max.value ? value : this.options.max.value;
-
-            const scaledValue = this.scale(value);
-            this.options.max.scaled = scaledValue >= this.options.max.scaled ? scaledValue : this.options.max.scaled;
 
             const percentage = this.percentage(value, total);
             const percentageLength = percentage ? percentage.toFixed(2).length + 5 : 0;
 
             const maxLabelLength = this.options.percentage ? label.length + percentageLength : label.length;
+            if (this.options.labels) this.options.max.label = Math.max(maxLabelLength, this.options.max.label);
 
-
-            if (this.options.labels) {
-                this.options.max.label = maxLabelLength > this.options.max.label ? maxLabelLength : this.options.max.label;
-            }
+            this.options.max.value = Math.max(value, this.options.max.value);
+            this.options.max.scaled = Math.max(this.scale(value), this.options.max.scaled);
 
             return a + value;
         }, 0);
@@ -54,7 +53,7 @@ class ChartProcessor {
 
     scale(value: number) {
         const size = this.options.orientation === 'vertical' ? this.options.height : this.options.width;
-        
+
         const { scale, max } = this.options;
 
         if (scale === "auto") {
@@ -82,15 +81,17 @@ class ChartProcessor {
         const chartData = new Map();
 
         processed.forEach((point, i) => {
-            const { color = this.options.color, label, value } = typeof point === "number" ? { value: point, label: point.toString() } : point;
+            const { color = this.options.color, label: pointLabel, value } = typeof point === "number" ? { value: point, label: point.toString() } : point;
             const scaled = Number(this.scale(value).toFixed(2));
+            const percentage = this.percentage(value, total);
+            const label = pointLabel || value.toString();
 
             const formattedPoint = {
-                label: label || value.toString(),
+                label,
                 value,
                 color,
                 scaled,
-                percentage: this.percentage(value, total)
+                percentage
             }
             chartData.set(i, formattedPoint);
         });
@@ -101,8 +102,8 @@ class ChartProcessor {
     sort(data: InputData[]): InputData[] {
         if (this.options.sort) {
             return data.sort((a, b) => {
-                const first = typeof a === "number" ? a : a.value;
-                const second = typeof b === "number" ? b : b.value;
+                const first = this.getPointValue(a);
+                const second = this.getPointValue(b);
                 return first - second;
             });
         }
