+++ types/types.ts
@@ -1,3 +1,5 @@
+
+
 export type Max = {
     label: number;
     value: number;
